Reduce some include of Cache.h since it needs 11s on average
